Updated crypto-functions and configuration handling.
